refactor: comprehensive code quality improvements and test fixes
  - Fix deprecated Gradle API usage: replace getTasks().create() with register().get()
    in TaskConfigurationServiceTest (6 occurrences)
  - Add try-with-resources blocks around Files.walk() calls to prevent resource leaks
    in CodegenConfigTemplateExtractor and OpenApiTemplateExtractor
  - Fix YAML syntax errors: resolve duplicate keys in enumClass.mustache.yaml and
    model.mustache.yaml template customization files
  - Improve documentation formatting: fix table layouts and convert numbered lists
    to auto-numbering format across all markdown files
  - Remove unused code: eliminate unused methods in CompiledSLF4JPattern,
    saveTemplate() in CodegenConfigTemplateExtractor, main() in OpenApiTemplateExtractor
  - Add comprehensive documentation to methodMissing() in OpenApiModelGenExtension
    explaining Groovy dynamic method dispatch for DSL configuration
  - Apply performance optimizations: substring→startsWith, configurations.all→configureEach
  - Fix test failures: directory creation assertion and library version compatibility
  - Replace console output with proper SLF4J logging in template extractors
  - Fix IntelliJ warnings: method return types, unused variables, deprecation warnings

  All 254 tests passing with full Gradle configuration cache compatibility maintained.

+++ plugin/src/main/java/com/guidedbyte/openapi/modelgen/constants/PluginConstants.java
@@ -84,7 +84,7 @@ public final class PluginConstants {
     /** Default hash display length for debugging (does not limit functionality). */
     public static final int HASH_DISPLAY_LENGTH = 16;
     /** Base retry delay in milliseconds for error handling. */
-    public static final int BASE_RETRY_DELAY_MS = 100;
+    public static final long BASE_RETRY_DELAY_MS = 100;
     
     // Task descriptions
     /** Description for setup template directories task. */
